requirements.txt
asgiref==3.7.2
asttokens==2.4.1
bleach==6.1.0
crispy-bootstrap5==2023.10
decorator==5.1.1
Django==5.0
django-autoslug==1.9.9
django-crispy-forms==2.1
django-extensions==3.2.3
django-summernote==0.8.20.0
django-tinymce==3.6.1
easy-thumbnails==2.8.5
executing==2.0.1
ipython==8.19.0
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-slugify==8.0.1
six==1.16.0
slugify==0.0.1
sqlparse==0.4.4
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.14.0
wcwidth==0.2.12
webencodings==0.5.1
